+++ src/shaders/PyMetalBridge.swift
@@ -47,12 +47,12 @@ public func init_shaders() {
 
 @available(macOS 10.13, *)
 @_cdecl("get_points_offsets")
-public func get_points_offsets(input: UnsafePointer<Int32>, out: UnsafeMutablePointer<Int32>, count: Int) -> Int {
+public func get_points_offsets(input: UnsafePointer<Int16>, out: UnsafeMutablePointer<Int16>, count: Int) -> Int {
     return computeOffsets(input: input, out: out, count: count)
 }
 
 @available(macOS 10.13, *)
-public func computeOffsets(input: UnsafePointer<Int32>, out: UnsafeMutablePointer<Int32>, count: Int) -> Int {
+public func computeOffsets(input: UnsafePointer<Int16>, out: UnsafeMutablePointer<Int16>, count: Int) -> Int {
     do {
         let inputBuffer = UnsafeRawPointer(input)
 
@@ -63,15 +63,15 @@ public func computeOffsets(input: UnsafePointer<Int32>, out: UnsafeMutablePointe
         let computePipelineState = try device.makeComputePipelineState(function: getPointsFunction)
         computeCommandEncoder!.setComputePipelineState(computePipelineState)
 
-        let inputByteLength = 5 * MemoryLayout<Int32>.size * count
+        let inputByteLength = 5 * MemoryLayout<Int16>.size * count
 
         let inVectorBuffer = device.makeBuffer(bytes: inputBuffer, length: inputByteLength, options: [])
 
         computeCommandEncoder!.setBuffer(inVectorBuffer, offset: 0, index: 0)
         computeCommandEncoder!.setBuffer(allTracksBuffer, offset: 0, index: 1)
 
-        let resultRef = UnsafeMutablePointer<Int32>.allocate(capacity: count)
-        let outVectorBuffer = device.makeBuffer(bytes: resultRef, length: count * MemoryLayout<Int32>.size, options: [])
+        let resultRef = UnsafeMutablePointer<Int16>.allocate(capacity: count)
+        let outVectorBuffer = device.makeBuffer(bytes: resultRef, length: count * MemoryLayout<Int16>.size, options: [])
 
         computeCommandEncoder!.setBuffer(outVectorBuffer, offset: 0, index: 2)
         
@@ -85,7 +85,7 @@ public func computeOffsets(input: UnsafePointer<Int32>, out: UnsafeMutablePointe
 
         // unsafe bitcast and assigin result pointer to output
 
-        out.initialize(from: outVectorBuffer!.contents().assumingMemoryBound(to: Int32.self), count: count)
+        out.initialize(from: outVectorBuffer!.contents().assumingMemoryBound(to: Int16.self), count: count)
 
         resultRef.deallocate()
