+++ src/shaders/Shaders.metal
@@ -2,16 +2,16 @@
 #include <metal_stdlib>
 using namespace metal;
 
-kernel void points_offsets(const device int *input [[ buffer(0) ]], const device uint8_t *track [[ buffer(1) ]], device int *out [[ buffer(2) ]], uint id [[ thread_position_in_grid ]]) {
-    int car_x = input[id * 5];
-    int car_y = input[id * 5 + 1];
-    int direction = input[id * 5 + 2];
-    int track_id = input[id * 5 + 3];
+kernel void points_offsets(const device short *input [[ buffer(0) ]], const device uint8_t *track [[ buffer(1) ]], device short *out [[ buffer(2) ]], uint id [[ thread_position_in_grid ]]) {
+    short car_x = input[id * 5];
+    short car_y = input[id * 5 + 1];
+    short direction = input[id * 5 + 2];
+    short track_id = input[id * 5 + 3];
     float ppm = (float)(input[id * 5 + 4]) / 1000; // pixels per meter (1000 = 1 meter
 
-    int x = car_x;
-    int y = car_y;
-    int distance = 0;
+    short x = car_x;
+    short y = car_y;
+    short distance = 0;
 
     if (track_id == -1) {
         out[id] = 0;
